requirements.txt
pocketflow>=0.0.1
pytesseract>=0.3.10
Pillow>=10.0.0